+++ languagetool-core/src/test/java/org/languagetool/tools/StringToolsTest.java
@@ -184,8 +184,8 @@ public class StringToolsTest extends TestCase {
     assertEquals(true, StringTools.isWhitespace("  "));
     assertEquals(true, StringTools.isWhitespace("\t"));
     assertEquals(true, StringTools.isWhitespace("\u2002"));
-    //non-breaking space is not a whitespace
-    assertEquals(false, StringTools.isWhitespace("\u00a0"));
+    //non-breaking space is also a whitespace
+    assertEquals(true, StringTools.isWhitespace("\u00a0"));
     assertEquals(false, StringTools.isWhitespace("abc"));
     //non-breaking OOo field
     assertEquals(false, StringTools.isWhitespace("\\u02"));